Simple solution for avoid deletions

+++ app/Http/Controllers/ThreadsController.php
@@ -118,6 +118,14 @@ class ThreadsController extends Controller
      */
     public function destroy($channel, Thread $thread)
     {
+        if($thread->user_id != auth()->id()){
+            if(request()->wantsJson()){
+                return response(['status' => 'Permission Denied'], 403);
+            }
+
+            return redirect('/login');
+        }
+        
         // 1.$thread->replies()->delete();
         // 2. Model events
         $thread->delete();
